pyproject.toml
[project]
name = "ipd-marl"
version = "0.1.0"
description = "Iterated Prisoner's Dilemma — Multi-Agent Reinforcement Learning experiments"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "numpy>=1.26",
    "pandas>=2.2",
    "matplotlib>=3.8",
    "axelrod>=4.12",
    "torch>=2.2",
    "seaborn>=0.13.2",
    "ruff>=0.15.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ipd_marl"]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
